move snackbar context to design system root
