+++ assets/src/edit-story/components/mediaPicker/test/useMediaPicker.js
@@ -23,7 +23,7 @@ import { act, renderHook } from '@testing-library/react-hooks';
  */
 import ConfigContext from '../../../app/config/context';
 import useMediaPicker from '../useMediaPicker';
-import SnackbarContext from '../../../../design-system/contexts/snackbar/context';
+import { SnackbarContext } from '../../../../design-system';
 
 jest.mock('../../../../design-system/contexts/snackbar/useSnackbar', () => ({
   useSnackbar: () => {
